move QuantumComponents
in helper/network/star_network/
